+++ pybop/costs/feature_distances.py
@@ -5,7 +5,9 @@ from scipy.optimize import minimize
 
 from pybop.costs.base_cost import BaseCost
 from pybop.costs.evaluation import Evaluation
+from pybop.parameters.parameter import Inputs
 from pybop.processing.dataset import Dataset
+from pybop.simulators.solution import Solution
 
 
 def indices_of(values, target):
@@ -75,6 +77,15 @@ class FeatureDistance(BaseCost):
                 self.target_data[self._target[0]][self.start_index : self.end_index],
             )
 
+    def evaluate(
+        self,
+        solution: Solution,
+        inputs: Inputs | None = None,
+        calculate_sensitivities: bool = False,
+    ) -> Evaluation:
+        """Evaluate the feature distance for the given solution."""
+        return Evaluation(self.__call__(y=solution[self.target[0]].data))
+
     def _inverse_fit_function(self, y, *args):
         return NotImplementedError
 
@@ -90,14 +101,19 @@ class FeatureDistance(BaseCost):
         involves applying the fit function to data and comparing to identity.
         """
         t = t - t[0]
-        fit_guess = self._fit_guess(t, y)
-        return self._feature_selection(
-            minimize(
-                lambda x: np.sum((t - self._inverse_fit_function(y, *x)) ** 2) ** 0.5,
-                x0=fit_guess,
-                method="trust-constr",
-            ).x
-        )
+        try:
+            fit_guess = self._fit_guess(t, y)
+            return self._feature_selection(
+                minimize(
+                    lambda x: (
+                        np.sum((t - self._inverse_fit_function(y, *x)) ** 2) ** 0.5
+                    ),
+                    x0=fit_guess,
+                    method="trust-constr",
+                ).x
+            )
+        except IndexError:
+            return self.failure(self.parameters.names, calculate_sensitivities=False)
 
     def __call__(
         self,
@@ -124,7 +140,7 @@ class FeatureDistance(BaseCost):
                     ]
                 )
             )
-        return Evaluation(error.item())
+        return error.item()
 
 
 class SquareRootFeatureDistance(FeatureDistance):