+++ pybop/models/lithium_ion/grouped_spm.py
@@ -495,6 +495,7 @@ class GroupedSPM(BaseGroupedModel):
         parameter_dictionary = {
             "Nominal cell capacity [A.h]": param["Nominal cell capacity [A.h]"],
             "Current function [A]": param["Current function [A]"],
+            "Ambient temperature [K]": param["Ambient temperature [K]"],
             "Initial temperature [K]": param["Ambient temperature [K]"],
             "Initial SoC": soc_init,
             "Minimum negative stoichiometry": x_0,